Actually create a crt signed by the Root CA!
Based on a dependency ordering specification, the crt file was not
actually getting signed by the root CA, but instead was self-signed.
While this works, I was trying to have the process demonstrate the
actual steps taken by a company (the server keystore operations) and the
CA (the root keystore operations).

After being more careful about the make target dependencies, it now does
what it should. Note: You do have to go in and explicitly trust the
server PEM in OSX!